requirements.txt
# All necessary dependencies will be installed from the robotics toolbox.
cbor2
coppeliasim-zmqremoteapi-client
ipython
matplotlib
numpy==1.26.4
pyzmq
roboticstoolbox-python==1.1.1
scipy
spatialmath-python
swift-sim
websockets<11
parameterized
